+++ viadot/sources/__init__.py
@@ -6,6 +6,7 @@ from .cloud_for_customers import CloudForCustomers
 from .sharepoint import Sharepoint
 from .bigquery import BigQuery
 from .salesforce import Salesforce
+from .databricks import Databricks
 
 
 try: